2010-05-28	Ralf Corsépius <ralf.corsepius@rtems.org>
	* umon/tfsDriver.c:
	Use size_t instead of int for pathnamelen.

+++ c/src/lib/libbsp/shared/umon/tfsDriver.c
@@ -241,7 +241,7 @@ static void fixPath(char *path)
 
 static int rtems_tfs_eval_path(
   const char                        *pathname,
-  int                                pathnamelen,
+  size_t                             pathnamelen,
   int                                flags,
   rtems_filesystem_location_info_t  *pathloc
 )
